requirements.txt
streamlit==1.12.2
numpy==1.23.2
pandas==1.4.4
Pillow==9.2.0

